+++ news-bundle/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Contao news bundle change log
 
 ### DEV
+
+ * Allow to sort news items by date, headline or randomly (see #13).